+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.47.0](https://github.com/searchspring/snap/compare/v0.46.0...v0.47.0) (2023-09-15)
+
+### Features
+
+- add facet filteredCount getter ([fb507b1](https://github.com/searchspring/snap/commit/fb507b11cdd6d1a825754094d184b289566e7c17))
+- add facet globals - limit and valueLimit to search response transform ([f86e04f](https://github.com/searchspring/snap/commit/f86e04f4b6367c4487c1e276fa2cfc002c7d56b9))
+- **controller/autocomplete:** changing keyup event to input for better support of copy/paste ([4a95251](https://github.com/searchspring/snap/commit/4a9525117af6bd376da952e0986ccf6efaaad4ce))
+
 # [0.46.0](https://github.com/searchspring/snap/compare/v0.45.1...v0.46.0) (2023-07-24)
 
 ### Features